refactor: moves LSP related code from plugin.ts into lsp.ts, decouples lsp from plugin.ts by removing internal 'plugins' reference and using a notification-listener pattern instead. feature: adds ability to specify a custom markdown renderer with a 'markdownRenderer' option
