+++ src/index.ts
@@ -1,10 +1,9 @@
 export {
-    LanguageServerClient,
     LanguageServerPlugin,
     languageServerWithClient,
     languageServer,
 } from "./plugin.js";
-
+export { LanguageServerClient } from "./lsp.js";
 export {
     languageId,
     documentUri,
